requirements.txt
numpy==1.20.1
opencv_python_headless==4.5.1.48
Pillow==9.3.0
pytest==7.1.3
scikit_image==0.19.3
setuptools==65.4.1
skimage==0.0
torch==1.7.0
torchvision==0.8.1
tqdm==4.64.1
